Date: 2023/07/17 Dev: ITA
Description:
1) ./src/components/GithubUsersList.js:
Added code to ensure that output is defined before making use of it.

2) ./src/hooks/useGetUsersFetch.js:
Initialised output to undefined, to make it easier to verify if it has realised the fetched data it was assigned.
Added default parameters to the custom hook and the requestFetch function.

3) ./src/components/SearchComponent.js
Added a useEffect and two useState hooks (submitted, fetched) to facilitate the fetching of data after the
user has typed and entered a search keyword.